Add Duration variants for ClientHttpRequestFactory timeouts
This commit adds overloaded variants of timeout setters that take a
Duration instead of a number.

+++ spring-web/src/main/java/org/springframework/http/client/JettyClientHttpRequest.java
@@ -46,10 +46,10 @@ class JettyClientHttpRequest extends AbstractStreamingClientHttpRequest {
 
 	private final Request request;
 
-	private final int readTimeout;
+	private final long readTimeout;
 
 
-	public JettyClientHttpRequest(Request request, int readTimeout) {
+	public JettyClientHttpRequest(Request request, long readTimeout) {
 		this.request = request;
 		this.readTimeout = readTimeout;
 	}
